Style improvements and all getter are const now.

+++ Color.cpp
@@ -19,7 +19,7 @@ Color::Color(void) : r{0.f}, g{0.f}, b{0.f} {}
 /*------------------------------------------------------------------------------
  * Constructor.
  */
-Color::Color(uint8_t _r, uint8_t _g, uint8_t _b) : r{_r}, g{_g}, b{_b} {}
+Color::Color(float _r, float _g, float _b) : r{_r}, g{_g}, b{_b} {}
 
 /*------------------------------------------------------------------------------
  * Function to define blue.
